fix: attempt to fix zeitwerk expectations and eager_load
