+++ lib/lutaml-xsd.rb
@@ -1,0 +1,1 @@
+require_relative "lutaml/xsd"
