+++ spec/lutaml/xsd_spec.rb
@@ -16,7 +16,9 @@ RSpec.describe Lutaml::Xsd do
   subject(:parsed_schema) { described_class.parse(schema, location: location) }
 
   Dir.glob(File.expand_path("fixtures/*.xsd", __dir__)).each do |input_file|
-    context "when parsing #{input_file}" do
+    rel_path = Pathname.new(input_file).relative_path_from(Pathname.new(__dir__)).to_s
+
+    context "when parsing #{rel_path}" do
       let(:schema) { File.read(input_file) }
       let(:location) { LOCATIONS[File.basename(input_file, ".xsd").to_sym] }
